add repositories, update models and update lombok
se agregan los contratos de repository e implementaciones en JDBC se corrigen algunos atributos en los modelos y se actualiza la versión de lombok para reconocer los getters y setters

+++ Laboratorio1/DeliveryAppBackend/src/main/java/com/app/DeliveryApp/models/Puntuacion.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Puntuacion {
     private Long idPuntuacion;
-    private int puntaje;
+    private Double puntaje;
     private String comentario;
 
     // FK
